+++ src/instances/garbage/instance.ts
@@ -1,7 +1,1 @@
-import type { WorkbenchInstanceRegistration } from '../instance-types'
-import { garbageInstance } from '../../../../apps/garbage/src/config.ts'
-
-export const workbenchInstanceRegistration = {
-  instance: garbageInstance,
-  default: true,
-} satisfies WorkbenchInstanceRegistration
+export { workbenchInstanceRegistration } from '../../../../apps/garbage/src/instance-registration.ts'